@@ -1,9 +1,5 @@
-#always building extensions in-place would break installation of the module distribution
-#[build_ext]
-#inplace=0
-
 [metadata]
-Metadata-Version = 2.1
+metadata_version = 2.1
 name = stopeight
 description = stopeight: Comparing sequences of points in 2 dimensions
 long_description = stopeight: Comparing sequences of points in 2 dimensions by visually overlapping them using matrix transformations (translation, scaling and rotation) and getting a boolean result.
@@ -29,4 +25,5 @@ classifiers =
 
 [options]
 zip_safe = False
-python_requires = >=3.5
+#extras_require = {"test": ["pytest>=6.0"]}
+python_requires = >=3.7
